split up index.js file
