+++ backend/prisma/prisma.js
@@ -1,0 +1,23 @@
+import { PrismaClient } from "@prisma/client";
+
+const globalForPrisma = globalThis;
+
+const prisma =
+  globalForPrisma.prisma ||
+  new PrismaClient({
+    datasources: {
+      db: {
+        url: `${process.env.DATABASE_URL}&sslmode=require`,
+      },
+    },
+    log:
+      process.env.NODE_ENV === "development"
+        ? ["query", "error", "warn"]
+        : ["error"],
+  });
+
+if (!globalForPrisma.prisma) {
+  globalForPrisma.prisma = prisma;
+}
+
+export default prisma;
